Purpose: enabled build with msvc 8.0 release config x86_64 both ansi and unicode

+++ trunk/dragndrop/dll/hldrapi.cpp
@@ -26,7 +26,10 @@ static MyStringW getHolderFileName()
     MyStringW res;
     res.length(chars);
 
-    while (chars == GetModuleFileName(reinterpret_cast<HMODULE>(getMyModuleBaseAddress()), res, chars))
+    while (chars == 
+            GetModuleFileName(
+                reinterpret_cast<HMODULE>(getMyModuleBaseAddress()), res, 
+                static_cast<DWORD>(chars)))
     {
         chars += MAX_PATH;
